pyproject.toml
[project]
name = "familly-nexus"
version = "0.1.0"
description = "a digital platform that allows family members to explore and share their collective memories"
authors = [
    {name = "jmarc580625", email = "jmarc580625+github@gmail.com"}
]
license = "MIT"
readme = "README.md"
requires-python = "^3.13"

[tool.poetry.dependencies]
Flask = "^3.1.0"
flask_sqlalchemy = "^3.0.0"
flask_cors = "^3.0.10"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
requests = "^2.32.3"
pgvector = "^0.3.6"
Pillow = "^10.0.0"  # For image processing and EXIF data extraction
boto3 = "^1.34.0"  # For AWS S3 integration
python-magic = "^0.4.27"  # For file type detection
alembic = "^1.11.0"  # For database migrations

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.1.2"
pytest = "^8.3.5"
pytest-cov = "^6.0.0"

[tool.pytest.ini_options]
pythonpath = ["backend"]
testpaths = ["backend/tests"]
addopts = "-v"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "backend" }]
